refactor(counter): move input to its own file
This was done in preparation that the Counter file will grow when I add
API calls, so its better to have the lower-level functionality of the
input in its own file.
